Refactor: BoardGameDto 통일
