+++ src/main/java/com/swyp/boardpick/dto/response/BoardGameDto.java
@@ -36,18 +36,6 @@ public class BoardGameDto {
     private List<UserBoardGame> userBoardGames;
     private List<String> tags;
 
-    public BoardGameDto(BoardGame boardGame, List<String> tags) {
-        this.id = boardGame.getId();
-        this.thumbnailUrl = boardGame.getThumbnailUrl();
-        this.imageUrl = boardGame.getImageUrl();
-        this.name = boardGame.getName();
-        this.description = boardGame.getDescription();
-        this.rating = boardGame.getRating();
-        this.ratingCount = boardGame.getRatingCount();
-        this.tags = tags;
-        this.picked = true;
-    }
-
     public BoardGameDto(BoardGame boardGame, String difficulty, List<String> boardGameCategories, List<String> tags, boolean picked) {
         this.id = boardGame.getId();
         this.thumbnailUrl = boardGame.getThumbnailUrl();
