+++ src/main/java/com/swyp/boardpick/service/implement/UserService.java
@@ -45,19 +45,6 @@ public class UserService {
         return null;
     }
 
-    public List<BoardGameDto> getMyPickList(Long id) {
-        return userRepository.findById(id)
-                .get().getUserBoardGames()
-                .stream().map(userBoardGame -> {
-                    BoardGame boardGame = userBoardGame.getBoardGame();
-                    List<String> tags = boardGame.getBoardGameTags()
-                            .stream().map(boardGameTag -> boardGameTag.getTag().getContent())
-                            .toList();
-                    return new BoardGameDto(userBoardGame.getBoardGame(), tags);
-                })
-                .toList();
-    }
-
     public UserDto getMyInfo(Long id) {
         User user =  userRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("User not found"));
